# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS API
  version: 0.1.0
  description: OpenAPI generado manualmente basado en comportamiento actual.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
      example:
        message: Registro no encontrado.
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
      required: [page, limit, total, total_pages]
      example:
        page: 1
        limit: 20
        total: 1
        total_pages: 1
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, example: admin@lms.local }
        first_name: { type: string, example: Admin }
        last_name: { type: string, example: LMS }
        roles:
          type: array
          items: { type: string }
          example: [ROLE_ADMIN]
        is_active: { type: boolean, example: true }
      required: [id, email, first_name, last_name, roles]
      example:
        id: 1
        email: admin@lms.local
        first_name: Admin
        last_name: LMS
        roles: [ROLE_ADMIN]
        is_active: true
...
    FileObject:
      type: object
      properties:
        id: { type: integer }
        key: { type: string }
        bucket: { type: string }
        original_name: { type: string }
        mime_type: { type: string }
        size: { type: integer }
      required: [id, key, bucket, original_name, mime_type, size]
      example:
        id: 10
        key: uploads/2026/02/abc.txt
        bucket: lms
        original_name: test.txt
        mime_type: text/plain
        size: 5
...
    ImportBatch:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        status: { type: string }
        total_rows: { type: integer, nullable: true }
        success_count: { type: integer, nullable: true }
        error_count: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        created_by:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            first_name: { type: string }
            last_name: { type: string }
        result_file:
          allOf:
            - $ref: '#/components/schemas/FileObject'
          nullable: true
      required: [id, type, status, created_at]
      example:
        id: 1
        type: users
        status: pending
        total_rows: null
        success_count: null
        error_count: null
        created_at: '2026-02-05 06:24:19'
        created_by:
          id: 1
          email: admin@lms.local
          first_name: Admin
          last_name: LMS
        result_file: null
paths:
...
  /users:
    get:
      tags: [Users]
      security: [bearerAuth: []]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: is_active
          schema: { type: boolean }
        - in: query
          name: role
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      tags: [Users]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                first_name: { type: string }
                last_name: { type: string }
                role: { type: string, enum: [ROLE_ADMIN, ROLE_TEACHER, ROLE_STUDENT] }
              required: [email, password, first_name, last_name]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/User' }
        '409':
          description: Conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
  /users/teachers:
    get:
      tags: [Users]
      security: [bearerAuth: []]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /users/admins:
    get:
      tags: [Users]
      security: [bearerAuth: []]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /users/{id}:
    put:
      tags: [Users]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                first_name: { type: string }
                last_name: { type: string }
                role: { type: string, enum: [ROLE_ADMIN, ROLE_TEACHER, ROLE_STUDENT] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/User' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
    delete:
      tags: [Users]
      security: [bearerAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...
  /imports/users:
    post:
      tags: [Imports]
      security: [bearerAuth: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file_id: { type: integer }
              required: [file_id]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/ImportBatch' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorMessage' }
...